hwdb: use /etc/udev/hwdb.bin for the binary database
This is a combination of ee623f0d0c04a9c8da4953ddc29862c8c4945458 and
1e1954f53386cb773e2a152748dd31c4d36aa2d8 from systemd.

Note that all Makefile changes from those two commits have been
excluded. This is because the build system has been rewritten and we do
not do anything with the hwdb right now. It still needs to be reviewed.

Signed-off-by: Richard Yao <ryao@gentoo.org>